+++ CHANGELOG.md
@@ -3,6 +3,233 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2023-10-19
+
+### Changes
+
+---
+
+Packages with breaking changes:
+
+ - [`studyu_flutter_common` - `v1.6.2`](#studyu_flutter_common---v162)
+
+Packages with other changes:
+
+ - [`studyu_core` - `v4.3.0`](#studyu_core---v430)
+ - [`studyu_app` - `v2.6.2`](#studyu_app---v262)
+ - [`studyu_designer_v2` - `v1.5.3`](#studyu_designer_v2---v153)
+
+Packages graduated to a stable release (see pre-releases prior to the stable version for changelog entries):
+
+ - `studyu_core` - `v4.3.0`
+ - `studyu_flutter_common` - `v1.6.2`
+
+---
+
+#### `studyu_flutter_common` - `v1.6.2`
+
+#### `studyu_core` - `v4.3.0`
+
+#### `studyu_app` - `v2.6.2`
+
+ - **FIX**: update descriptions and rename imprint. ([6d3757f4](https://github.com/hpi-studyu/studyu/commit/6d3757f425dd5b95ee581e10b6ad675f8e71786a))
+
+#### `studyu_designer_v2` - `v1.5.3`
+
+ - **FIX**: update descriptions and rename imprint. ([6d3757f4](https://github.com/hpi-studyu/studyu/commit/6d3757f425dd5b95ee581e10b6ad675f8e71786a))
+
+
+## 2023-10-11
+
+### Changes
+
+---
+
+Packages with breaking changes:
+
+ - There are no breaking changes in this release.
+
+Packages with other changes:
+
+ - [`studyu_app` - `v2.6.1`](#studyu_app---v261)
+ - [`studyu_core` - `v4.3.0-dev.10`](#studyu_core---v430-dev10)
+ - [`studyu_designer_v2` - `v1.5.2`](#studyu_designer_v2---v152)
+ - [`studyu_flutter_common` - `v1.6.2-dev.11`](#studyu_flutter_common---v162-dev11)
+
+---
+
+#### `studyu_app` - `v2.6.1`
+
+ - **FIX**: upgrade deps. ([6fa0e8b8](https://github.com/hpi-studyu/studyu/commit/6fa0e8b8332c0872b59c7c31042fb845f0708e73))
+ - **FIX**: update cocoapod. ([ed45f6f5](https://github.com/hpi-studyu/studyu/commit/ed45f6f5a71e8227652d291232769bcc0bb75ba6))
+
+#### `studyu_core` - `v4.3.0-dev.10`
+
+ - **FIX**: upgrade deps. ([6fa0e8b8](https://github.com/hpi-studyu/studyu/commit/6fa0e8b8332c0872b59c7c31042fb845f0708e73))
+
+#### `studyu_designer_v2` - `v1.5.2`
+
+ - **FIX**: upgrade deps. ([6fa0e8b8](https://github.com/hpi-studyu/studyu/commit/6fa0e8b8332c0872b59c7c31042fb845f0708e73))
+
+#### `studyu_flutter_common` - `v1.6.2-dev.11`
+
+ - **FIX**: upgrade deps. ([6fa0e8b8](https://github.com/hpi-studyu/studyu/commit/6fa0e8b8332c0872b59c7c31042fb845f0708e73))
+ - **FEAT**: switch to hpi env. ([7f1a1817](https://github.com/hpi-studyu/studyu/commit/7f1a1817d96d955cf2f9e7c7fb7467f317bec53b))
+
+
+## 2023-09-27
+
+### Changes
+
+---
+
+Packages with breaking changes:
+
+ - There are no breaking changes in this release.
+
+Packages with other changes:
+
+ - [`studyu_designer_v2` - `v1.5.1`](#studyu_designer_v2---v151)
+
+---
+
+#### `studyu_designer_v2` - `v1.5.1`
+
+ - **FIX**: rename custom MustMatchValidator. ([613e14a4](https://github.com/hpi-studyu/studyu/commit/613e14a4c846cae5a4ba3f83294a70978d907ca1))
+
+
+## 2023-09-27
+
+### Changes
+
+---
+
+Packages with breaking changes:
+
+ - There are no breaking changes in this release.
+
+Packages with other changes:
+
+ - [`studyu_app` - `v2.6.0`](#studyu_app---v260)
+ - [`studyu_core` - `v4.3.0-dev.9`](#studyu_core---v430-dev9)
+ - [`studyu_designer_v2` - `v1.5.0`](#studyu_designer_v2---v150)
+ - [`studyu_flutter_common` - `v1.6.2-dev.10`](#studyu_flutter_common---v162-dev10)
+
+---
+
+#### `studyu_app` - `v2.6.0`
+
+ - **REFACTOR**: hardcoded baseline id. ([aeac14a2](https://github.com/hpi-studyu/studyu/commit/aeac14a2e2b01c0d48799d9ed69264d72f170353))
+ - **REFACTOR**: bar generator. ([a05c7808](https://github.com/hpi-studyu/studyu/commit/a05c78082828551381483c15434cddf25ab681ac))
+ - **REFACTOR**: remove code duplication. ([a3ff4203](https://github.com/hpi-studyu/studyu/commit/a3ff4203f883170e0730e79d5b0aa677e0f1aa8a))
+ - **REFACTOR**: remove unnecessary passing of instance variable. ([5142f0e8](https://github.com/hpi-studyu/studyu/commit/5142f0e8a088acacfd881d4d1dbcf0c01d5a603b))
+ - **REFACTOR**: pass diagram data. ([7062d4f0](https://github.com/hpi-studyu/studyu/commit/7062d4f034d8fa0a24886bed7ab12525d674ea39))
+ - **FIX**: upgrade deps. ([7cdc7c55](https://github.com/hpi-studyu/studyu/commit/7cdc7c559f64cfdd63ee07efbeeb028c418f6f4e))
+ - **FIX**: upgrade deps. ([17213746](https://github.com/hpi-studyu/studyu/commit/172137469ed236dc694979bbfe017708faabbd77))
+ - **FIX**: revert minSdkVersion to 19. ([c928cf9b](https://github.com/hpi-studyu/studyu/commit/c928cf9b893455d8cc9ce436799f7a83930f742c))
+ - **FIX**: reports for studies without baseline. ([231915ae](https://github.com/hpi-studyu/studyu/commit/231915aef5f7d22e5749600b779177ba36b7f9e8))
+ - **FIX**: day offset when baseline is excluded. ([5b3577aa](https://github.com/hpi-studyu/studyu/commit/5b3577aa381b2d554c11a88d24798c82b361a7ca))
+ - **FIX**: chart import. ([7d616580](https://github.com/hpi-studyu/studyu/commit/7d616580ccc11473117a255f0f94ee08432db639))
+ - **FIX**: bump minimum required sdk version to 33 and set target to flutter.compileSdkVersion. ([92bec536](https://github.com/hpi-studyu/studyu/commit/92bec536c676b1ff361030e4b045428c9c1b025d))
+ - **FEAT**: add 20% y-padding. ([0d5694d6](https://github.com/hpi-studyu/studyu/commit/0d5694d61e74f368a450fc1cba075f9599713dfd))
+ - **FEAT**: bold title. ([e2c9914e](https://github.com/hpi-studyu/studyu/commit/e2c9914ec1243ebb89b7218f975faa3948646a7b))
+ - **FEAT**: dynamic bar width. ([474f9223](https://github.com/hpi-studyu/studyu/commit/474f922325f42c525691f0c2e0b84446181190af))
+ - **FEAT**: display phase number as integer. ([0dab0269](https://github.com/hpi-studyu/studyu/commit/0dab026951c13c13b7f8c96f85c5c1109ca40fcb))
+ - **FEAT**: enable tooltips. ([3b642f5b](https://github.com/hpi-studyu/studyu/commit/3b642f5b99376dad324da41a9b69470f36470c90))
+ - **FEAT**: phase separator lines. ([5b6921a5](https://github.com/hpi-studyu/studyu/commit/5b6921a565ec7f8c3fb1cdbdfd84e2a8e8dfb9b1))
+ - **FEAT**: add question title. ([0450f925](https://github.com/hpi-studyu/studyu/commit/0450f925bbe388d4dc4eb8a49061b93e8b92587c))
+ - **FEAT**: chart legend. ([4a5984e2](https://github.com/hpi-studyu/studyu/commit/4a5984e29220cc1ce88ba981d6ba2679681fe2b0))
+ - **FEAT**: get intervention names. ([8525cc7d](https://github.com/hpi-studyu/studyu/commit/8525cc7d838ec2c4a97d4454747b56e474c45c5f))
+ - **FEAT**: legend widgets. ([8031b88e](https://github.com/hpi-studyu/studyu/commit/8031b88e491152ed4c37b83b34ec2b8f5b2af051))
+
+#### `studyu_core` - `v4.3.0-dev.9`
+
+ - **FIX**: upgrade deps. ([7cdc7c55](https://github.com/hpi-studyu/studyu/commit/7cdc7c559f64cfdd63ee07efbeeb028c418f6f4e))
+ - **FIX**: upgrade deps. ([17213746](https://github.com/hpi-studyu/studyu/commit/172137469ed236dc694979bbfe017708faabbd77))
+
+#### `studyu_designer_v2` - `v1.5.0`
+
+ - **FIX**: upgrade deps. ([7cdc7c55](https://github.com/hpi-studyu/studyu/commit/7cdc7c559f64cfdd63ee07efbeeb028c418f6f4e))
+ - **FIX**: upgrade deps. ([17213746](https://github.com/hpi-studyu/studyu/commit/172137469ed236dc694979bbfe017708faabbd77))
+ - **FIX**: report button visible for read-only studies. ([99daed13](https://github.com/hpi-studyu/studyu/commit/99daed1356840eceb640755a06daa967a8c28fb1))
+ - **FIX**: regenerate ids when duplicating a study. ([79b411c3](https://github.com/hpi-studyu/studyu/commit/79b411c3b4dfe9467c412a7cd2c47508ce766623))
+ - **FEAT**: localization additions. ([787f2ba6](https://github.com/hpi-studyu/studyu/commit/787f2ba659a39427ae988181ef89f20f24cdb121))
+
+#### `studyu_flutter_common` - `v1.6.2-dev.10`
+
+ - **FIX**: upgrade deps. ([7cdc7c55](https://github.com/hpi-studyu/studyu/commit/7cdc7c559f64cfdd63ee07efbeeb028c418f6f4e))
+ - **FIX**: upgrade deps. ([17213746](https://github.com/hpi-studyu/studyu/commit/172137469ed236dc694979bbfe017708faabbd77))
+
+
+## 2023-08-16
+
+### Changes
+
+---
+
+Packages with breaking changes:
+
+ - There are no breaking changes in this release.
+
+Packages with other changes:
+
+ - [`studyu_app` - `v2.5.0`](#studyu_app---v250)
+ - [`studyu_core` - `v4.3.0-dev.8`](#studyu_core---v430-dev8)
+ - [`studyu_designer_v2` - `v1.4.0`](#studyu_designer_v2---v140)
+ - [`studyu_flutter_common` - `v1.6.2-dev.9`](#studyu_flutter_common---v162-dev9)
+
+---
+
+#### `studyu_app` - `v2.5.0`
+
+ - **FIX**: make analytics fully optional. ([67d58887](https://github.com/hpi-studyu/studyu/commit/67d588879a69a9105cd2e54c08f2a2feae1224a2))
+ - **FIX**: upgrade deps. ([aa2cb251](https://github.com/hpi-studyu/studyu/commit/aa2cb2512e3db4e31f97edea8d2bccf55e91f9d2))
+ - **FEAT**: make study tags reactive. ([f2f8ecaf](https://github.com/hpi-studyu/studyu/commit/f2f8ecafb4f21d21128ce089eb99608778b91e2e))
+
+#### `studyu_core` - `v4.3.0-dev.8`
+
+ - **FIX**: make analytics fully optional. ([67d58887](https://github.com/hpi-studyu/studyu/commit/67d588879a69a9105cd2e54c08f2a2feae1224a2))
+ - **FIX**: upgrade deps. ([aa2cb251](https://github.com/hpi-studyu/studyu/commit/aa2cb2512e3db4e31f97edea8d2bccf55e91f9d2))
+ - **FIX**: study was not returned completely. ([324e2ceb](https://github.com/hpi-studyu/studyu/commit/324e2ceb2afe25df5ee06cd2450865137efe7ae1))
+ - **FIX**: make Repo urls optional. ([d0a96ea8](https://github.com/hpi-studyu/studyu/commit/d0a96ea8413bed3e86a4448da6f06b35b4b9adca))
+ - **FIX**: tag color. ([9baddd1d](https://github.com/hpi-studyu/studyu/commit/9baddd1d3df22ebb45fa5117f724fd9b98b31b98))
+ - **FIX**: tag related sql. ([59882b7e](https://github.com/hpi-studyu/studyu/commit/59882b7e86b7145131733c4d6851cf621b676fc8))
+ - **FIX**: study_tags was null for new studies. ([db57e09f](https://github.com/hpi-studyu/studyu/commit/db57e09f841164b39d587c5ba6479a1f7a73b57e))
+ - **FIX**: make color nullable and int, add tags to dashboard. ([ef7532a3](https://github.com/hpi-studyu/studyu/commit/ef7532a3e571abd140cfbeee1e2e1b6fd14f9ed6))
+ - **FIX**: tag type. ([cf39195b](https://github.com/hpi-studyu/studyu/commit/cf39195b540fe6b65e4d600586041b3c19d90d52))
+ - **FEAT**: working tags with foreign key study_tag. ([8d4eaa92](https://github.com/hpi-studyu/studyu/commit/8d4eaa925c8a921bfa70f0afae6a57a155a55c9e))
+ - **FEAT**: display and update tags correctly. ([b96dcf8a](https://github.com/hpi-studyu/studyu/commit/b96dcf8a25def52dfc59bc65206177e0ab4a25d8))
+ - **FEAT**: initial study tags. ([615d5b62](https://github.com/hpi-studyu/studyu/commit/615d5b629a2ae5d9998d50b1a0f119fa62a2bc87))
+
+#### `studyu_designer_v2` - `v1.4.0`
+
+ - **REFACTOR**: format. ([595a717f](https://github.com/hpi-studyu/studyu/commit/595a717f0e09a3ba311984ad1f4b584252e091ec))
+ - **REFACTOR**: pin changes to user_repository.dart. ([e6579c8b](https://github.com/hpi-studyu/studyu/commit/e6579c8bc9f6b9c5e828c81fe7d185ccbfa7f8d0))
+ - **REFACTOR**: cleanup. ([b4a987e7](https://github.com/hpi-studyu/studyu/commit/b4a987e7add7100a23f2366a83d6551b6d861567))
+ - **REFACTOR**: rename. ([c1bf13dd](https://github.com/hpi-studyu/studyu/commit/c1bf13dd8f0f8d1ea4ea3ee8fe4663c0700d486a))
+ - **REFACTOR**: make tags more reactive. ([d7cb5f30](https://github.com/hpi-studyu/studyu/commit/d7cb5f30f10834a3bf90c2b74e400cd7103846e0))
+ - **REFACTOR**: fix dashboard pin null value. ([eda8ccfb](https://github.com/hpi-studyu/studyu/commit/eda8ccfb3a0da3a1d6a2eb986c85bd24526aea0d))
+ - **FIX**: upgrade deps. ([aa2cb251](https://github.com/hpi-studyu/studyu/commit/aa2cb2512e3db4e31f97edea8d2bccf55e91f9d2))
+ - **FIX**: study was not returned completely. ([324e2ceb](https://github.com/hpi-studyu/studyu/commit/324e2ceb2afe25df5ee06cd2450865137efe7ae1))
+ - **FIX**: non-working pin. ([5729a096](https://github.com/hpi-studyu/studyu/commit/5729a096d599827026b68cfce910e6a78e309a35))
+ - **FIX**: tag color. ([9baddd1d](https://github.com/hpi-studyu/studyu/commit/9baddd1d3df22ebb45fa5117f724fd9b98b31b98))
+ - **FIX**: tag related sql. ([59882b7e](https://github.com/hpi-studyu/studyu/commit/59882b7e86b7145131733c4d6851cf621b676fc8))
+ - **FIX**: make color nullable and int, add tags to dashboard. ([ef7532a3](https://github.com/hpi-studyu/studyu/commit/ef7532a3e571abd140cfbeee1e2e1b6fd14f9ed6))
+ - **FIX**: update copyright. ([515785bd](https://github.com/hpi-studyu/studyu/commit/515785bd4717693d5cb65c7fd79a1c70ecf8c441))
+ - **FIX**: remove Designer V2 version number from page title. ([362a4fe5](https://github.com/hpi-studyu/studyu/commit/362a4fe5f5a31b4364d9695f8f2b7193d3ac811e))
+ - **FEAT**: make study tags reactive. ([f2f8ecaf](https://github.com/hpi-studyu/studyu/commit/f2f8ecafb4f21d21128ce089eb99608778b91e2e))
+ - **FEAT**: refactor Search to widget and make tags clickable. ([95f33064](https://github.com/hpi-studyu/studyu/commit/95f33064444b9e594d9497bc6321c529253ee0bd))
+ - **FEAT**: working tags with foreign key study_tag. ([8d4eaa92](https://github.com/hpi-studyu/studyu/commit/8d4eaa925c8a921bfa70f0afae6a57a155a55c9e))
+ - **FEAT**: display and update tags correctly. ([b96dcf8a](https://github.com/hpi-studyu/studyu/commit/b96dcf8a25def52dfc59bc65206177e0ab4a25d8))
+ - **FEAT**: initial study tags. ([615d5b62](https://github.com/hpi-studyu/studyu/commit/615d5b629a2ae5d9998d50b1a0f119fa62a2bc87))
+
+#### `studyu_flutter_common` - `v1.6.2-dev.9`
+
+ - **REFACTOR**: rename .env.selfhost to .env.hpi. ([dae7d461](https://github.com/hpi-studyu/studyu/commit/dae7d4617a3a45df595172672851c0f61ff14de5))
+ - **REFACTOR**: rename .env.selfhost to .env.hpi. ([0c3d2252](https://github.com/hpi-studyu/studyu/commit/0c3d2252edaf7f564a417e5b8f4243161514cc37))
+ - **FIX**: upgrade deps. ([aa2cb251](https://github.com/hpi-studyu/studyu/commit/aa2cb2512e3db4e31f97edea8d2bccf55e91f9d2))
+
+
 ## 2023-07-05
 
 ### Changes